requirements.txt
bagit==1.7.0
CacheControl==0.11.7
certifi==2019.9.11
chardet==3.0.4
coloredlogs==10.0
cwl-utils==0.1
cwlref-runner==1.0
cwltool==1.0.20190915164430
decorator==4.4.0
future==0.17.1
humanfriendly==4.18
idna==2.8
isodate==0.6.0
lockfile==0.12.2
lxml==4.4.1
mistune==0.8.4
mypy-extensions==0.4.1
networkx==2.3
pathlib2==2.3.5
prov==1.5.1
psutil==5.6.3
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.22.0
ruamel.yaml==0.15.97
ruamel.yaml.clib==0.1.2
scandir==1.10.0
schema-salad==4.5.20190906201758
shellescape==3.4.1
six==1.12.0
typing-extensions==3.7.4
urllib3==1.25.6
